+++ merged-packages/json-rpc-engine/package.json
@@ -4,7 +4,7 @@
   "description": "a tool for processing JSON RPC",
   "main": "index.js",
   "scripts": {
-    "test": "node test.js"
+    "test": "node test/basic.js"
   },
   "author": "",
   "license": "ISC",
