+++ merged-packages/json-rpc-engine/test/basic.js
@@ -1,4 +1,4 @@
-const RpcEngine = require('./index.js')
+const RpcEngine = require('../index.js')
 const test = require('tape')
 
 test('basic middleware test', function(t){
@@ -146,33 +146,6 @@ test('return order of events', function(t){
     t.equals(events[2], '3-end', '(event 2) was "3-end"')
     t.equals(events[3], '2-return', '(event 3) was "2-return"')
     t.equals(events[4], '1-return', '(event 4) was "1-return"')
-    t.end()
-  })
-
-})
-
-test('test asMiddleware', function(t){
-
-  let engine = new RpcEngine()
-  let subengine = new RpcEngine()
-  let originalReq = undefined
-
-  subengine.push(function(req, res, next, end){
-    originalReq = req
-    res.result = 'saw subengine'
-    end()
-  })
-
-  engine.push(subengine.asMiddleware())
-
-  let payload = { id: 1, jsonrpc: '2.0', method: 'hello' }
-
-  engine.handle(payload, function(err, res){
-    t.ifError(err, 'did not error')
-    t.ok(res, 'has res')
-    t.equals(originalReq.id, res.id, 'id matches')
-    t.equals(originalReq.jsonrpc, res.jsonrpc, 'jsonrpc version matches')
-    t.equals(res.result, 'saw subengine', 'response was handled by nested engine')
     t.end()
   })
