+++ merged-packages/json-rpc-engine/test/index.js
@@ -1,0 +1,2 @@
+require('./basic.js')
+require('./asMiddleware.js')
